Add menu and initialize backing data for terminal
As soon as I wanted to `printf()' in the dependency board.c, I knew it
would be leaky (abstraction). Instead, send in the _representation_ of a
display, and let main sort it out. After all, main is where stdio.h is
included.

Stubbed out a display of a board, since there will be logic there based
on adjacency.

Took longer than I thought to "print a string." First, I tried to be
clever and start from (len-1), modifying len. But that prevents easily
tacking on the null-terminating character later.

I know I'll need a MAKEFILE sooner than later.