Fix logic for printing board across multiple rows
It's simpler than what I thought before, needing lookahead. We just need
to print a newline under these situations (of adjacents):

A. After each row in the middle (neither first nor last row)

    ... 8 5

B. After first row, upper-right corner

    ... 5 3

C. On last row, lower-right corner

    ... 5 3

To do that, we use two variables: one to record the current element as
the last element seen before the next loop, and another to count the
first row, increment (but skip adding an newline) to distinguish between
5s between corner 3s & 5s enveloping 8s, and then reset it for the
last row.

Note 5 and then 3 happens twice: in first row, and second-to-last row
then last row (last 5, first 3). As data is a string, the sequence is
the same but we want to _interpret_ it differently.