Display last column number to right of each row
On the board, we show the last index for each row. The user can then put
a number between that and some smaller number to designate a position on
a particular row. It would be difficult otherwise to choose where to put
the first letter of a word to add.

Where to put the info was easy enough: just before printing a newline.
How and what to print took a bit longer.

We have to fill in an array representing a display one character at a
time; for boards with more than 9 positions, we need to calculate each
digit. So position 18 would mean adding a `1' and an `8' next to each
other to display that.

Textbooks have solutions for this, but wanted to see if I could write my
own. Exponentiation could also come from math.h, but I wanted to try a
"multiply as a series of adds" to explore the link between summation and
multiplication.

The row_number() function shows some of my initial attempt. It took a
couple days of thinking off-and-on to break it into two concepts: base
ten processing for one, and peeling off digits one at a time as the
other.

Next commit, I'll remove the diagnostic and exploratory function. Just
wanted a record of what I had tried in the history.