Include design & intuition of offset as direction
We use the distance between index 0 (first corner) to the next corner
(the northeast corner) as a "unit ruler," an offset that gives us some
useful properties to calculate top-down direction (like traveling on a
2d plane).

This helps us develop intuition between 2d space and offset, useful for
more than just a word search. (Memory management, hopefully.)

Code changes are more a commit of design notes therein. Cleanup will
likely remove the bounds_left and bounds_right calculations in favor of
something that works with the unit ruler.

As well, wanted to commit something before it gets too big and too much
is rewritten that could have been useful notes on the journey.