Calculate directions with unit ruler and start pos
Checkpoint commit, before I get too far in the weeds. The unit ruler is
shaping up to be useful! I'm able to calculate index positions for all
eight directions. Now just need to filter it out.

We'll need to run it each time for each letter in the input word, and
consult states[] for whether there is room. That's another thing to work
on.

Advent of Code 2024 also started...