Calculate min and max extents per board row
The extents define the row's minimum and maximum index. We collect this
to determine whether a east or west direction is allowed.

While iterating the loop, we determine the maximum rows. We must cap
extents, because a board can have less "rows" than "directions." That
is, as numbers, extents calculated on directions' eight indices results
in numbers larger than the board!

Next, we'll refactor to a function and work on sanity checks based on
the direction. For example, east at the end of a row cannot be allowed.