Separate usage of data index as board index
Checkpoint commit. We were iterating on directions[], but we were also
calculating `max_extents_index' per row. For positions like nw (0),
n (1), and ne (2), which are on the "top row" of our clockwise compass,
the board row is "zero" but the index `i' is 0, 1, and 2.

We actually just need to calculate the "current row" based on `start',
and use that throughout to check the eight directions. It's a subtle bug
but we did find it.

This is the perfect example of small changes that balloon the diff, that
would motivate small commits. But it's hard to pin down C methods as we
are noodling away. With unit testing, we'd have to tweak the tests too.
(But that would help solidify things.)