Refactor once we know extents indices are constant
It looks like we only need min and max extents per iteration based on
the board row index, so we can remove some extra bookkeeping code.

In order to reset directions back to -1, we have to accept bad input
like negative index calculations _into_ the reset method. So we can't
protect ourselves with assertions for valid inputs in that case.

Is there another approach? Should we front a purer method with an
invalid input receiver? May not be worth the cost of an extra function
that lets us early return--plus having assertions as needed.