Fix board row index off-by-one with early aborting
We have to be cautious after using a variable that is more than a clever
counter. If the variable's value of 0 means something, we may need to
`break' to avoid (bug) breakage.

Test cases so far are using YOLO at positions 5 and 11. Need to think on
whether it's possible to infer ne positions are invalid with given input
or if we need more data to reset for sure.

Needing only the bare minimum to conclude something is its own
cleverness, I guess. At least we removed two params earlier, since we
only needed to calculate max and min extents per row. Before, it was a
flag and a frontier.